Refactor to fit Antenna structure and satisfy PMD
