+++ modules/disclosure-document/disclosure-document-core/src/main/java/org/eclipse/sw360/antenna/disclosure/document/core/ArtifactAndLicense.java
@@ -42,7 +42,7 @@ public interface ArtifactAndLicense {
     */
    Optional<String> getCopyrightStatement();
 
-   public class LicenseInfo {
+   class LicenseInfo {
 
       private final String key, title, text, shortName;
