+++ modules/disclosure-document/disclosure-document-generator/src/test/java/org/eclipse/sw360/antenna/disclosure/document/generator/ArtifactAdapterTest.java
@@ -14,6 +14,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.io.File;
 
+import org.eclipse.sw360.antenna.disclosure.document.workflow.generators.ArtifactAdapter;
 import org.eclipse.sw360.antenna.model.artifact.Artifact;
 import org.eclipse.sw360.antenna.model.artifact.facts.ArtifactFile;
 import org.eclipse.sw360.antenna.model.artifact.facts.java.MavenCoordinates;
