+++ modules/disclosure-document/disclosure-document-generator/src/test/java/org/eclipse/sw360/antenna/disclosure/document/generator/ArtifactAndLicenseComparatorTest.java
@@ -12,6 +12,7 @@ package org.eclipse.sw360.antenna.disclosure.document.generator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.eclipse.sw360.antenna.disclosure.document.workflow.generators.ArtifactAndLicenseComparator;
 import org.junit.jupiter.api.Test;
 
 /**
